requirements.txt
selenium==4.30.0
pytest==8.3.5
pytest-html==4.1.1
python-dotenv==1.1.0
allure-pytest==2.13.5
webdriver-manager==4.0.2
pytest-bdd==8.1.0
PyYAML==6.0.1
Pillow==10.0.0